pyproject.toml
[project]
name = "api-backend"
version = "0.1.0"
description = "Sustainable Market Backend API"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "alembic>=1.15.2",
    "authlib>=1.5.2",
    "flask>=3.0.2",
    "flask-cors>=4.0.0",
    "flask-limiter>=3.5.0",
    "flask-migrate>=4.1.0",
    "flask-sqlalchemy>=3.1.1",
    "flask-swagger-ui>=4.11.1",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.10",
    "pyjwt>=2.8.0",
    "pytest>=8.3.5",
    "pytest-cov>=6.1.1",
    "python-dotenv>=1.1.0",
    "requests>=2.31.0",
    "supabase>=2.15.0",
    "flask-jwt-extended>=4.6.0",
    "flask-compress>=1.14",
    "flask-talisman>=1.1.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
filterwarnings = ["ignore::DeprecationWarning"]
pythonpath = ["src"]

[tool.coverage.run]
source = ["src"]
omit = ["tests/*", "migrations/*"]
